app1Cryptography/microservice1/requirements.txt
uvicorn
fastapi
pydantic
aiohttp
asyncio
pycryptodome